requirements.txt
pyTelegramBotAPI==3.7.1
psycopg2-binary
peewee
python-decouple